+++ skyfield_data/__init__.py
@@ -4,7 +4,7 @@ from .expirations import check_expirations
 __DATA_PATH = abspath(join(dirname(__file__), "data"))
 
 
-def get_skyfield_data_path():
+def get_skyfield_data_path(expiration_limit=0):
     """
     Return the data path to be used along with Skyfield loader.
 
@@ -14,8 +14,12 @@ def get_skyfield_data_path():
         from skyfield.api import Loader
         load = Loader(get_skyfield_data_path())
         planets = load('de421.bsp')
+
+    :param expiration_limit: Limit in days for expiration calculation.
+                             Default is zero (only check for expired files)
+    :type expiration_limit: int
     """
-    check_expirations()
+    check_expirations(expiration_limit)
     return __DATA_PATH